\name{confint.lnre}
\alias{confint.lnre}
\title{Confidence intervals for LNRE model parameters (zipfR)}

\description{

  \bold{TODO:} write documentation for \code{confint.lnre} method

}

\usage{
\method{confint}{lnre}(object, parm, level=0.95, method=c("normal", "mad", "empirical"), plot=FALSE, \dots)
}

\arguments{

  \item{object}{\bold{TODO}}

  \item{parm}{\bold{TODO}}

  \item{level}{\bold{TODO}}

  \item{method}{\bold{TODO}}

  \item{plot}{\bold{TODO}}

  \item{\dots}{\bold{TODO}}

}

\details{

  \bold{TODO}

}

\seealso{

  \bold{TODO}

}

\examples{

  ## TODO 

}

\keyword{ htest }
